Some manual merging with memory_gradient branch. After solving the rotation bug, it's time for new irace trials with a good initial condition

+++ src/rw_gradient_follower.py
@@ -18,26 +18,43 @@ import shutil
 
 home_path = os.path.expanduser('~')
 
+# TODO: pass the gradient_memory flag from the config
+gradient_memory = False
+
 
 def main():
     random.seed(argv[1])
-    config_file = generate_config_file(argv[1:])
-    config = Configuration(config_file=config_file)
+    if os.path.exists(argv[2]):
+        # print("Config file exists")
+        config_file = 'None'
+        config = Configuration(config_file=argv[2])
+    else:
+        # print("Config file NOT exists")
+        config_file = generate_config_file(argv[1:])
+        config = Configuration(config_file=config_file)
+
     if config.parameters["VISUALIZE"] != 0:
         main_controller = MainController(config)
         _ = ViewController(main_controller,
                            config.parameters["WIDTH"],
                            config.parameters["HEIGHT"],
                            config.parameters["FPS"])
+
+        overall_gradient = main_controller.get_overall_gradient() / \
+                            config.parameters['NB_ROBOTS'] / \
+                            config.parameters['SIMULATION_STEPS']
+        print(f"Overall gradient {overall_gradient}")
     else:
-        run(config)                 # this is for single run
-        os.remove(config_file)
+        run(config)  # this is for single run
+        if os.path.exists(config_file):
+            os.remove(config_file)
 
 
 def generate_config_file(list_args):
     # print('list_args: ', list_args)
     config_path = list_args[3]
-    irace_config = home_path+"/swarm_aggregation/config/irace/irace_config"+list_args[0]+'_'+list_args[1]+'_'+list_args[2]+".txt"
+    irace_config = home_path + "/swarm_aggregation/config/irace/irace_config" + list_args[0] + '_' + list_args[
+        1] + '_' + list_args[2] + ".txt"
     # print('list_args: ', list_args)
     shutil.copyfile(config_path, irace_config)
     # print(list_args[2:])
@@ -51,16 +68,17 @@ def generate_config_file(list_args):
         exit(1)
 
     with open(irace_config, "a") as file:
-        q_bits_str = 'QUANTIZATION_BITS='+q_bits
+        q_bits_str = 'QUANTIZATION_BITS=' + q_bits
         std_str = 'STD_MOTION_STEPS='
-        for i in range(int(q_bits)):
+        _values = int(q_bits) * int(q_bits) if gradient_memory else int(q_bits)
+        for i in range(_values):
             std_str += std + ','
 
         file.write(q_bits_str + '\n')
         file.write(std_str[:-1] + '\n')
 
     reshape_args = np.array(list_args[7:]).reshape(-1, 2)
-    num_param = reshape_args.shape[0]//int(q_bits)
+    num_param = reshape_args.shape[0] // _values
     for i in range(num_param):
         idxs = np.arange(i, reshape_args.shape[0], num_param)
         values = np.take(reshape_args[:, 1], idxs)
@@ -72,8 +90,8 @@ def generate_config_file(list_args):
         if 'st' in reshape_args[i, 0]:
             param = 'STD_MOTION_STEPS='
 
-        array_str = param+str(values).replace(' [', '').replace('[', '').replace(']', '').replace(' ', ',')\
-                                                                                         .replace("'", "")
+        array_str = param + str(values).replace(' [', '').replace('[', '').replace(']', '').replace(' ', ',') \
+            .replace("'", "")
 
         with open(irace_config, "a") as file:
             file.write(array_str + '\n')
@@ -110,13 +128,16 @@ def run(config):
     gradientPOS_file = f"center_gradient_positions.txt"
     filepath = f"../data"
     os.makedirs(filepath, exist_ok=True)
+    overall_gradient = controller.get_overall_gradient() / \
+                       config.parameters['NB_ROBOTS'] / \
+                       config.parameters['SIMULATION_STEPS']
     with open(f"{filepath}/{gradientVAL_file}", "a") as file:
-        file.write(f"{argv[1:]}: {controller.get_overall_gradient()}\n")
+        file.write(f"{argv[1:]}: {overall_gradient}\n")
     with open(f"{filepath}/{gradientPOS_file}", "a") as file:
         file.write(f"{argv[1:]}: {controller.environment.center_gradient}\n")
     # print(f"Overall gradient for {config.parameters['NB_ROBOTS']} robots: "
     #       f"{controller.get_overall_gradient()}")
-    print(f"{controller.get_overall_gradient()}")
+    print(f"{overall_gradient}")
 
 
 if __name__ == '__main__':
